Fix merging from upstream and relink moved files ! Tests failing
