+++ karma.conf.js
@@ -25,7 +25,7 @@ module.exports = function(config) {
         // list of files / patterns to load in the browser
         files: [
             'node_modules/bignumber.js/bignumber.js',
-            'test/*.js',
+            'test/**/*.js',
         ],
