+++ lib/web3.js
@@ -23,19 +23,19 @@
  * @date 2014
  */
 
-if (process.env.NODE_ENV !== 'build') {
-    var BigNumber = require('bignumber.js');
-}
+// if (process.env.NODE_ENV !== 'build') {
+//     var BigNumber = require('bignumber.js');
+// }
 
 var net            = require('./web3/net');
-var eth            = require('./eth');
-var db             = require('./db');
-var shh            = require('./shh');
-var watches        = require('./watches');
-var filter         = require('./filter');
-var utils          = require('./utils');
+var eth            = require('./web3/eth');
+var db             = require('./web3/db');
+var shh            = require('./web3/shh');
+var watches        = require('./web3/watches');
+var filter         = require('./web3/filter');
+var utils          = require('./utils/utils');
 var formatters     = require('./solidity/formatters');
-var requestManager = require('./requestmanager');
+var requestManager = require('./web3/requestmanager');
 var c              = require('./utils/config');
 
 /// @returns an array of objects describing web3 api methods
